requirements.txt
Django==3.2.5
channels==3.0.4
pyjwt==2.3.0